+++ src/components/BlockchainList/EtherScanLink.tsx
@@ -1,5 +1,5 @@
 import { LinkText, StatusText } from 'components/ui/Text'
-import getEtherscanTxUrl from 'helpers/getEtherscanTxUrl'
+import getEtherscanTxUrl from 'helpers/network/getEtherscanTxUrl'
 
 export default function ({ tx }: { tx: string }) {
   return (
